// app/lib/generated/yield-bonds/src/generated/types/redemptionType.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/codecs";

/** Describes the origin of a pending redemption request. */
export enum RedemptionType {
  BondSale,
  PrizeClaim,
  FeeWithdrawal,
}

export type RedemptionTypeArgs = RedemptionType;

export function getRedemptionTypeEncoder(): FixedSizeEncoder<RedemptionTypeArgs> {
  return getEnumEncoder(RedemptionType);
}

export function getRedemptionTypeDecoder(): FixedSizeDecoder<RedemptionType> {
  return getEnumDecoder(RedemptionType);
}

export function getRedemptionTypeCodec(): FixedSizeCodec<
  RedemptionTypeArgs,
  RedemptionType
> {
  return combineCodec(getRedemptionTypeEncoder(), getRedemptionTypeDecoder());
}