// app/lib/generated/yield-bonds/src/generated/types/redemptionBatchStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/codecs";

export enum RedemptionBatchStatus {
  Accumulating,
  Submitted,
  Settled,
}

export type RedemptionBatchStatusArgs = RedemptionBatchStatus;

export function getRedemptionBatchStatusEncoder(): FixedSizeEncoder<RedemptionBatchStatusArgs> {
  return getEnumEncoder(RedemptionBatchStatus);
}

export function getRedemptionBatchStatusDecoder(): FixedSizeDecoder<RedemptionBatchStatus> {
  return getEnumDecoder(RedemptionBatchStatus);
}

export function getRedemptionBatchStatusCodec(): FixedSizeCodec<
  RedemptionBatchStatusArgs,
  RedemptionBatchStatus
> {
  return combineCodec(
    getRedemptionBatchStatusEncoder(),
    getRedemptionBatchStatusDecoder()
  );
}